refactor: success페이지 파일명 수정 및 뒤로가기 네비게이트 추가
